fix wrong audio timeCode
