+++ src/components/sentence/SentenceCard.vue
@@ -12,11 +12,6 @@
       :can-redo="canRedo"
     ></SentenceToolBar>
 
-    <AudioPlayer v-if="isSound()"
-     :sentence-data="sentenceData"
-     :reactive-sentences-obj="(reactiveSentencesObj as reactive_sentences_obj_t)"
-    ></AudioPlayer>
-
     <div>
       <q-tabs
         v-model="openTabUser"
@@ -201,7 +196,6 @@ import VueDepTree from './VueDepTree.vue';
 import XposDialog from './XposDialog.vue';
 import RelationDialog from './RelationDialog.vue';
 import SentenceToolBar from './SentenceToolBar.vue';
-import AudioPlayer from './AudioPlayer.vue';
 
 
 function sentenceBusFactory(): sentence_bus_t {
@@ -225,7 +219,6 @@ export default defineComponent({
     MultiEditDialog,
     SentenceToolBar,
     RelationDialog,
-    AudioPlayer,
   },
   props: {
     index: {
@@ -518,11 +511,6 @@ export default defineComponent({
     },
     removeUserTag(tag: string) {
       this.removeTag(this.sentenceData, tag, this.sentenceBus, this.openTabUser);
-    },
-    isSound() {
-      const [conllData] = Object.values(this.sentenceData.conlls)
-      const soundUrl= conllData.match(/sound_url = (.*?)\n/)
-      return soundUrl !== null
     },
   },
 });
